requirements/test.txt
-r common.txt
django-coverage
django-nose
mock
nosexcover
-e git+git@github.com:SeanHayes/django-inmemorystorage.git#egg=django-inmemorystorage
